Updated search by sku for variants and products
