+++ src/lib/types/product.ts
@@ -1,0 +1,17 @@
+export type VariantType = {
+  node: {
+    sku: string;
+    id: string;
+    title: string;
+  };
+};
+
+export type ProductType = {
+  node: {
+    id: string;
+    title: string;
+    variants: {
+      edges: VariantType[];
+    };
+  };
+};
